Store the Etherpad API key as a secret.
Currently, the Etherpad API key were public, with the rationale the
installation is a public one, everybody can anonymously or with a name
make edits, destroy pads, create new pads, etc.

Nasqueron is interested to get the potential to integrate other applications
with Etherpad, using groups, sessions and users objects to make pads spawned
by 3rd party application private.

A private API key has so been regenerated, and we now store it as a secret.