Adding Composer in the mix
We want to introduce Apache log4php to offer logging capabilities, for
example when our Etherpad API key doesn't match the one installed on
the server.

API applications are now invited to include _includes/core.php so they'll
get access to Composer autoload facilities, and in the future, standard
features (logging is coming in the next commit).

A Makefile is shipped to call composer update, .gitignore is introduced
to ignore composers volatile files and folders.