Send a 500 HTTP response code on Etherpad API issue
Currently, if the Etherpad API key isn't correct, or if the server
can't be reached, the server returns a HTTP 200 reply, with an empty
array.

This change keeps the empty array reply, but with a 500 HTTP code
to indicate there is currently an issue and so pads can't be fetched.